Test Results and Bench Mark
As shown above, the iterative Fibonacci series run much faster than the recursive Fibonacci series, and this is because all the factions must be called, stored, and then returned to the calling faction. It is very simple when it comes to coding compared to iteration, and its simplicity is the trade-off for its speed. On the other hand, the iteration that mainly depends on looping is a complex system but much faster than the recursive. According to the results, recursion appears to use more memory, but only by using larger numbers can you tell the difference in memory usage. 